Moves ComplianceRequestPage out of Compliance directory
